backend3/requirements.txt
# backend3/requirements.txt
# Core
langchain
langgraph
langchain-openai
langchain-community

# Tools & Data
pypdf
chromadb
python-dotenv
duckduckgo-search
lxml

# Web Server
flask
flask-cors